+++ src/MicroElements.Metadata/Metadata/PropertyExtensions.cs
@@ -69,30 +69,18 @@ namespace MicroElements.Metadata
         /// <typeparam name="B">Result type.</typeparam>
         /// <param name="property">Source property.</param>
         /// <param name="map">Function that maps value of type <typeparamref name="A"/> to type <typeparamref name="B"/>.</param>
-        /// <param name="allowMapNull">By default false: does not calls <paramref name="map"/> if <paramref name="property"/> value is null.</param>
         /// <param name="searchOptions">Search options to get property value.</param>
         /// <returns>New property of type <typeparamref name="B"/>.</returns>
         public static IProperty<B> Map<A, B>(
             this IProperty<A> property,
             Func<IPropertyValue<A>, (B Value, ValueSource ValueSource)> map,
-            bool allowMapNull = false,
             SearchOptions searchOptions = default)
         {
             (B, ValueSource) ConvertValue(IPropertyContainer container)
             {
                 var propertyValueA = container.GetPropertyValue(property, searchOptions);
-                if (propertyValueA.HasValue())
-                {
-                    A valueA = propertyValueA.Value;
-                    bool shouldMap = !valueA.IsNull() || (valueA.IsNull() && allowMapNull);
-                    if (shouldMap)
-                    {
-                        var valueBResult = map(propertyValueA);
-                        return valueBResult;
-                    }
-                }
-
-                return (default, ValueSource.NotDefined);
+                var valueBResult = map(propertyValueA);
+                return valueBResult;
             }
 
             return new Property<B>(property.Name).SetCalculate(container => ConvertValue(container));
@@ -125,7 +113,27 @@ namespace MicroElements.Metadata
                 allowMapNull: false,
                 searchOptions
                     .UseDefaultValue(false)
-                    .CalculateValue(true));
+                    .CalculateValue(true))
+                .WithAlias($"{property.Name}_Nullified");
+        }
+
+        /// <summary>
+        /// Calculates undefined struct value to <paramref name="defaultValue"/>.
+        /// </summary>
+        /// <typeparam name="A">Value type.</typeparam>
+        /// <param name="property">Source property.</param>
+        /// <param name="defaultValue">Default value.</param>
+        /// <param name="searchOptions">Search options to get property value.</param>
+        /// <returns>New property of type <typeparamref name="A"/>.</returns>
+        public static IProperty<A> UseDefaultForUndefined<A>(
+            this IProperty<A> property,
+            A defaultValue = default,
+            SearchOptions searchOptions = default)
+            where A : struct
+        {
+            return property
+                .Map(pv => pv.IsNullOrNotDefined() ? (defaultValue, ValueSource.Calculated) : (pv.Value, pv.Source), searchOptions: searchOptions)
+                .WithAlias($"{property.Name}_WithDefaultForUndefined");
         }
     }
 }